bugs fixed: 1. all prices types are set to be BigDecimal types and their jdbc types are DECIMAL 2. all date type deserialized as @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8") 3. adding seckill will callback its seckillId 4. deleting seckill will not remove it from db, instead of modifying its isDeleted to 1 5. when seckillList is empty, it will return empty list correctly

+++ src/main/java/com/tsong/cmall/controller/vo/OrderVO.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -21,7 +22,7 @@ public class OrderVO implements Serializable {
     private String orderNo;
 
     @Schema(title = "订单总价")
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 
     @Schema(title = "支付类型")
     private Byte payType;
